More samples, including some multi-type models
Not all of these models are working yet. Includes documentation on known errors
with possible fixes.

+++ tflitehub/test_util.py
@@ -12,6 +12,7 @@ import iree.compiler.tflite as iree_tflite_compile
 import iree.runtime as iree_rt
 import numpy as np
 import os
+import shutil
 import sys
 import tensorflow.compat.v2 as tf
 import urllib.request
@@ -28,8 +29,11 @@ class TFLiteModelTest(testing.absltest.TestCase):
     os.makedirs(self.workdir, exist_ok=True)
     self.tflite_file = '/'.join([self.workdir, 'model.tflite'])
     self.tflite_ir = '/'.join([self.workdir, 'tflite.mlir'])
-
-    urllib.request.urlretrieve(self.model_path, self.tflite_file)
+    self.iree_ir = '/'.join([self.workdir, 'tosa.mlir'])
+    if os.path.exists(self.model_path):
+      shutil.copy(self.model_path, self.tflite_file)
+    else:
+      urllib.request.urlretrieve(self.model_path, self.tflite_file)
     self.binary = '/'.join([self.workdir, 'module.bytecode'])
 
   def generate_inputs(self, input_details):
@@ -61,6 +65,7 @@ class TFLiteModelTest(testing.absltest.TestCase):
       self.tflite_file, input_type="tosa",
       output_file=self.binary,
       save_temp_tfl_input=self.tflite_ir,
+      save_temp_iree_input=self.iree_ir,
       target_backends=iree_tflite_compile.DEFAULT_TESTING_BACKENDS,
       import_only=False)
